Add cat Calls and Fix Comments for wrap_contents.sh
A trio cat calls have been added to wrap_contents.sh to wrap the
contents of a file between the header and footer HTML files that were
specified. In addition, the variable comments have been fixed to reflect
the true intent of the variables. There were a couple of
misunderstandings that confused their purposes.